fix: postDetail page refactor
